+++ EngineClasses.cpp
@@ -10,207 +10,10 @@
 *   Moreover, it would be better to create main parent class - ResurseHolder (for music, images, fonts and etc.)
 */
 
-settings mysetts;
+extern settings mysetts;
+extern FontHolder font_holder;
 sf::View g_view;
 TextureHolder texture_holder;
-
-//____________________________________________________________________
-
-FontHolder::FontHolder() { }
-
-void FontHolder::load(Fonts::ID id, const std::string& filename)
-{
-	std::unique_ptr<sf::Font> font(new sf::Font());
-	font->loadFromFile(filename);
-
-	gFontMap.insert(std::make_pair(id, std::move(font)));
-}
-
-sf::Font& FontHolder::get(Fonts::ID id)
-{
-	auto found = gFontMap.find(id);
-	return *found->second;
-}
-
-const sf::Font& FontHolder::get(Fonts::ID id) const
-{
-	auto found = gFontMap.find(id);
-	return *found->second;
-}
-
-// in future make by singleton/think about better decision
-FontHolder font_holder;
-
-//____________________________________________________________________
-
-
-NickName::NickName()
-{
-	text.setFont(font_holder.get(Fonts::OLD));
-	text.setFillColor(sf::Color::Red);
-	text.setCharacterSize(10);
-}
-
-void NickName::set_coordinates(float x, float y)
-{
-	int x_offset = 0;
-
-	if (nick_length <= 9) x_offset += 10 - nick_length + 3;
-	else x_offset -= (nick_length % 10) / 2;
-	text.setPosition(x + x_offset, y - 10);
-	// std::cout << std::endl << x << " " << y << std::endl;
-}
-
-void NickName::set_string(std::string nick)
-{
-	text.setString(nick);
-	nick_length = nick.length();
-}
-
-void NickName::drawU(sf::RenderWindow& window)
-{
-	window.draw(text);
-}
-
-//____________________________________________________________________
-
-UserInput::UserInput()
-{ 
-	text.setFont(font_holder.get(Fonts::OLD));
-	text.setFillColor(sf::Color::Red);
-	text.setString("____________________");
-	inp = "____________________";
-	text.setPosition(mysetts.get_width()/2.5f - 80, mysetts.get_height()/2);
-}
-
-void UserInput::set_size(int size)
-{
-	text.setCharacterSize(size);
-}
-
-void UserInput::set_coordinates(float x, float y)
-{
-	text.setPosition(x, y);
-}
-
-std::string UserInput::inputting(sf::RenderWindow& window)
-{
-	sf::Font& font = font_holder.get(Fonts::OLD);
-	sf::Text text1("", font, 20);
-	//text.setOutlineColor(sf::Color::Red);
-	text1.setFillColor(sf::Color::Red);
-	text1.setStyle(sf::Text::Bold);
-	text1.setString("Input your nickname:");
-	text1.setPosition(mysetts.get_width() / 3.f, mysetts.get_height() / 2.5f);
-
-	while (window.isOpen())
-	{
-		window.clear();
-		window.draw(text);
-		window.draw(text1);
-		window.display();
-
-		bool enter = false;
-		sf::Event event;
-		while (window.pollEvent(event))        // SFML is saving all events in a queue, so checking them all
-		{
-			sf::Keyboard::Key key;
-			char symb = ' ';
-
-			bool backspace = false;
-			bool smth = false;             // if it's needed to be added
-			switch (event.type)
-			{
-			case sf::Event::KeyPressed:
-				if (event.key.code == sf::Keyboard::Enter)
-				{
-					enter = true;
-					break;
-				}
-
-				else if (event.key.code == sf::Keyboard::BackSpace)
-				{
-					backspace = true;
-					break;
-				}
-				
-				// THEIR ENUM STARTS FROM UNKNOWN = -1, A = 0 (-> B = 1). AAAAAAAAAAAAA. 
-				// It can be useful to read docs smts...
-				else if (static_cast<int>(event.key.code) >= 0 && static_cast<int>(event.key.code) <= static_cast<int>(sf::Keyboard::Z))
-				{
-					symb = static_cast<int>(event.key.code) + 97;
-					smth = true;
-					//std::cout << symb;
-				}
-
-				else if (static_cast<int>(event.key.code) >= 26 && static_cast<int>(event.key.code) <= static_cast<int>(sf::Keyboard::Num9))
-				{
-					symb = static_cast<int>(event.key.code) + 22;
-					smth = true;
-					//std::cout << symb;
-				}
-			break;
-
-			case  sf::Event::Closed:
-				window.close();
-				break;
-			}
-			if (enter) break;
-
-			if (backspace)
-			{
-				for (int i = inp.length() - 1; i >= 0; --i)
-				{
-					if (inp[i] != '_')
-					{
-						inp[i] = '_';
-						break;
-					}
-				}
-				backspace = false;
-			}
-
-			if (smth)
-			{
-				for (int i = 0; i < inp.length(); ++i)
-				{
-					if (inp[i] == '_')
-					{
-						inp[i] = symb;
-						break;
-					}
-				}
-				smth = false;
-			}
-		}
-		if (enter) break;
-		text.setString(inp);
-
-	}
-
-	std::string output{};
-	for (int i = 0; i < inp.length(); ++i)
-	{
-		if (inp[i] != '_') output += inp[i];
-		else break;
-	}
-		
-	return output;
-}
-
-std::string UserInput::get_input()
-{
-	std::string output{};
-	for (int i = 0; i < inp.length(); ++i)
-	{
-		if (inp[i] != '_') output += inp[i];
-		else break;
-	}
-
-	return output;
-}
-
-//____________________________________________________________________
 
 void set_view(float x, float y)
 {
@@ -243,266 +46,7 @@ void set_view(float x, float y)
 
 //____________________________________________________________________
 
-Textures::ID InvItem::get_id()
-{
-	return id;
-}
 
-InvItem::InvItem(Textures::ID id): id(id)
-{
-	amount = 0;
-
-	sf::Texture& texture = texture_holder.get(id);
-	sprite.setTexture(texture);
-	sprite.setScale(0.3, 0.3);
-}
-
-int InvItem::get_amount()
-{
-	return amount;
-}
-
-void InvItem::add_one()
-{
-	amount++;
-}
-
-void InvItem::set_position(sf::Vector2f pos)
-{
-	sprite.setPosition(pos);
-}
-
-void InvItem::drawU(sf::RenderWindow& window)
-{
-	window.draw(sprite);
-}
-
-void InvItem::substract_one()
-{
-	amount--;
-}
-
-void InvItem::add_plenty(int num)
-{
-	amount += num;
-}
-
-void InvItem::set_scale(sf::Vector2f scale)
-{
-	sprite.setScale(scale);
-}
-
-sf::Vector2f InvItem::get_position()
-{
-	return sprite.getPosition();
-}
-
-InvItem::InvItem(Textures::ID id, int kolvo): id(id)
-{
-	amount = kolvo;
-
-	sf::Texture& texture = texture_holder.get(id);
-	sprite.setTexture(texture);
-	sprite.setScale(0.3, 0.3);
-}
-
-
-//____________________________________________________________________
-
-void Inventory::key_reaction(sf::Keyboard::Key key)
-{
-	bool first_current = false;
-	if (current_item <= 0 || current_item > 8)
-	{
-		current_item = key;
-		first_current = true;
-	}
-
-	if (first_current)
-	{
-		current_item = static_cast<int>(key) - 26;
-		cubes[current_item - 1].setOutlineColor(sf::Color::Red);
-	}
-
-	else
-	{
-		int prev_cur = current_item;
-		cubes[current_item - 1].setOutlineColor(sf::Color::Green);
-		current_item = static_cast<int>(key) - 26;
-		cubes[current_item - 1].setOutlineColor(sf::Color::Red);
-
-		int i = 1;
-		for (auto it = items.begin(); it != items.end(); it++)
-		{
-			if (i == prev_cur)
-			{
-				(*it).set_scale(sf::Vector2f(0.3, 0.3));
-				break;
-			}
-			++i;
-		}
-	}
-
-
-	int i = 1;
-	for (auto it = items.begin(); it != items.end(); it++)
-	{
-		if (i == current_item)
-		{
-			(*it).set_scale(sf::Vector2f(0.5, 0.5));
-			break;
-		}
-		++i;
-	}
-}
-
-void Inventory::update_statement()
-{
-	int scr_x = g_view.getCenter().x - mysetts.get_width() / 2;
-	int scr_y = g_view.getCenter().y - mysetts.get_height() / 2;
-
-	int pos_x = 164;
-	int pos_y = 448;
- 	for (int i = 0; i < 8; ++i)
-	{
-		cubes[i].setPosition(sf::Vector2f(scr_x + pos_x, scr_y + pos_y));
-		pos_x += 40;
-	}
-
-	inv_line.setPosition(sf::Vector2f(scr_x + 120, scr_y + 440));
-
-	for (auto it = items.begin(); it != items.end(); it++)
-		if ((*it).get_amount() == 0)
-			items.erase(it);
-
-	int i = 0;
-	for (auto it = items.begin(); it != items.end(); it++)
-	{
-		if (i == current_item-1)
-			(*it).set_position(cubes[i].getPosition() + sf::Vector2f(4.f, 4.f));
-		else
-		(*it).set_position(cubes[i].getPosition() + sf::Vector2f(10.f, 10.f));
-
-		++i;
-		if (i == 8) break;
-	}
-}
-
-void Inventory::decrease_item()
-{
-	if (current_item > items.size()) return;
-
-	int i = 1;
-	for (auto it = items.begin(); it != items.end(); it++)
-	{
-		if (i == current_item)
-		{
-			if ((*it).get_amount() == 1)
-				items.erase(it);
-
-			else
-				(*it).substract_one();
-			return;
-		}
-		++i;
-	}
-}
-
-int Inventory::set_current(int num)
-{
-	return 1;
-}
-
-Textures::ID Inventory::get_current()
-{
-	if (current_item > items.size() || current_item <= 0)
-		return Textures::ID::NUL;
-	
-	int i = 1;
-	for (auto it = items.begin(); it != items.end(); it++)
-	{
-		if (i == current_item)
-		{
-			return (*it).get_id();
-		}
-		++i;
-	}
-	return Textures::ID::DIRT;
-
-}
-
-Inventory::Inventory()
-{
-	for (int i = 0; i < 8; ++i)
-	{
-		cubes[i].setSize(sf::Vector2f(32.f, 32.f));
-		cubes[i].setFillColor(sf::Color(128, 128, 128));
-		cubes[i].setOutlineThickness(3.f);
-		cubes[i].setOutlineColor(sf::Color::Green);
-
-	}
-	// cubes[0].setOutlineColor(sf::Color::Red);
-
-	inv_line.setSize(sf::Vector2f(400.f, 40.f));
-	inv_line.setFillColor(sf::Color::Yellow);
-
-}
-
-void Inventory::drawU(sf::RenderWindow& window)
-{
-	window.draw(inv_line);
-	for (auto it = cubes.begin(); it != cubes.end(); it++)
-		window.draw(*it);
-
-	int i = 0;
-	for (auto it = items.begin(); it != items.end(); it++)
-	{
-		std::string kol = std::to_string((*it).get_amount());
-		sf::Text text(kol, font_holder.get(Fonts::ID::OLD), 10);
-		text.setFillColor(sf::Color::Black);
-		
-		if (i == current_item - 1) 
-			text.setPosition((*it).get_position() + sf::Vector2f(16.f, -12.f));
-		else
-			text.setPosition((*it).get_position() + sf::Vector2f(10.f, -12.f));
-		((*it).drawU(window));
-		window.draw(text);
-		i++;
-;		if (i == 8) break;
-	}
-
-}
-
-void Inventory::add_item(Textures::ID id, int kolvo)
-{
-	if (kolvo <= 0) return;
-	bool added_to_end = false;
-	auto iter = items.end();
-	for (auto it = items.begin(); it != items.end(); it++)
-		if ((*it).get_id() == id)
-			iter = it;
-	if (iter != items.end())
-		(*iter).add_plenty(kolvo);
-	else
-	{
-		items.push_back(InvItem(id, 0));
-		added_to_end = true;
-	}
-
-	if (added_to_end)
-	{
-		int i = 0;
-		for (auto it = items.begin(); it != items.end(); it++)
-		{
-			if (i == items.size() - 1)
-			{
-				(*it).add_plenty(kolvo);
-			}
-			++i;
-		}
-	}
-	// std::cout << "amount" << (*(items.begin())).get_amount();
-}
 
 //____________________________________________________________________
