+++ EngineClasses.h
@@ -28,12 +28,7 @@ TextureHolder modifiing:
 */
 
 #pragma once
-#include <cmath>
-#include <vector>
-#include <string>
-#include <memory>
-#include "WorldClass.h"
-
+#include "InventoryClasses.h"
 
 // a little bit of shitcode (let it be, it's a joke) for fixed time steps realisation - my proud
 // it gives about 60 fps, but work (may be) more stable, then standart function
@@ -42,95 +37,8 @@ const sf::Time TimePerFrame = sf::seconds(1.f / 60*1.f);
 
 void set_view(float x, float y);
 
-class NickName
-{
-private:
-	sf::Text text;
-	int nick_length = 0;
 
-public:
-	NickName();
-	void set_coordinates(float x, float y);
-	void set_string(std::string nick);
-	void drawU(sf::RenderWindow& window);
 
-};
-
-class UserInput
-{
-private:
-	std::string inp;
-	sf::Text text;
-
-public:
-	UserInput();
-	void set_size(int size);
-	void set_coordinates(float x, float y);
-	std::string inputting(sf::RenderWindow& window);
-	std::string get_input();
-
-};
-
-namespace Fonts
-{
-	enum ID { OLD };
-}
-
-class FontHolder
-{
-private:
-	std::map<Fonts::ID, std::unique_ptr<sf::Font>> gFontMap;
-
-public:
-	FontHolder();
-	void load(Fonts::ID id, const std::string& filename);
-	sf::Font& get(Fonts::ID id);
-	const sf::Font& get(Fonts::ID id) const;
-};
-
-class InvItem
-{
-public:
-	Textures::ID get_id();
-	int get_amount();
-	InvItem(Textures::ID id);
-	InvItem(Textures::ID id, int kolvo);
-	void add_one();
-	void add_plenty(int num);
-	void set_scale(sf::Vector2f scale);
-    sf::Vector2f get_position();
-	void substract_one();
-	void set_position(sf::Vector2f pos);
-	void drawU(sf::RenderWindow& window);
-	int set_item_type(Textures::ID id);
-
-private:
-	int item_type;
-	Textures::ID id;
-	sf::Sprite sprite;
-	int amount;
-
-};
-
-class Inventory
-{
-public:
-	void drawU(sf::RenderWindow& window);
-	Inventory();
-	void update_statement();
-	void add_item(Textures::ID id, int kolvo = 1);
-	Textures::ID get_current();
-	void key_reaction(sf::Keyboard::Key key);
-	int set_current(int num);
-	void decrease_item();
-
-private:
-	std::list<InvItem> items;
-	std::array<sf::RectangleShape, 8> cubes;
-	sf::RectangleShape inv_line;
-	int current_item = 0;
-
-};
 
 class Player
 {
