+++ main.cpp
@@ -1,4 +1,4 @@
-#include "mainfiles/EngineClasses.h"
+#include "EngineClasses.h"
 
 int main()
 {
